LayoutTests/http/tests/from-origin/document-nested-from-origin-same-blocked-expected.txt

CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
Tests that a nested cross-origin document load fails if the server blocks cross-origin loads with a 'From-Origin: same' response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
The outer iframe, loading a nested localhost iframe.



--------
Frame: '<!--frame2-->'
--------
